requirements.txt
tweepy>=4.0.0
beautifulsoup4>=4.0.0
requests>=2.0.0
python-dotenv>=0.0.0
sqlite3>=3.0.0
pytest>=7.0.0
python-dateutil>=2.8.0